Loss objects now in own .py
