+++ projetos/ReCycleGAN/src/models/__init__.py
@@ -1,1 +1,2 @@
 from .cyclegan import CycleGAN
+from .losses import Loss, LossLists, LossValues
